bot/requirements.txt
python-telegram-bot[webhooks]>=21.0
redis>=5.0.3
asyncpg>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
